// frontend/web/src/services/API/proserve-wb-projects-api/models/GetProjectAssignmentsResponseItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Projects API
 * API in charge of processing the methods required for the use cases
 *
 * The version of the OpenAPI document: 2024-04-11T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { GetProjectAssignmentsResponseItemADGroup } from './GetProjectAssignmentsResponseItemADGroup';
import {
    GetProjectAssignmentsResponseItemADGroupFromJSON,
    GetProjectAssignmentsResponseItemADGroupFromJSONTyped,
    GetProjectAssignmentsResponseItemADGroupToJSON,
} from './GetProjectAssignmentsResponseItemADGroup';

/**
 * 
 * @export
 * @interface GetProjectAssignmentsResponseItem
 */
export interface GetProjectAssignmentsResponseItem {
    /**
     * User TID
     * @type {string}
     * @memberof GetProjectAssignmentsResponseItem
     */
    userId?: string;
    /**
     * User roles in the project
     * @type {Array<string>}
     * @memberof GetProjectAssignmentsResponseItem
     */
    roles?: Array<string>;
    /**
     * Email address of the user
     * @type {string}
     * @memberof GetProjectAssignmentsResponseItem
     */
    userEmail?: string;
    /**
     * Active directory groups that the user is member of
     * @type {Array<GetProjectAssignmentsResponseItemADGroup>}
     * @memberof GetProjectAssignmentsResponseItem
     */
    activeDirectoryGroups?: Array<GetProjectAssignmentsResponseItemADGroup>;
    /**
     * Status of adding the user to the AD groups
     * @type {string}
     * @memberof GetProjectAssignmentsResponseItem
     */
    activeDirectoryGroupStatus?: string;
}

/**
 * Check if a given object implements the GetProjectAssignmentsResponseItem interface.
 */
export function instanceOfGetProjectAssignmentsResponseItem(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GetProjectAssignmentsResponseItemFromJSON(json: any): GetProjectAssignmentsResponseItem {
    return GetProjectAssignmentsResponseItemFromJSONTyped(json, false);
}

export function GetProjectAssignmentsResponseItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetProjectAssignmentsResponseItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'userId': !exists(json, 'userId') ? undefined : json['userId'],
        'roles': !exists(json, 'roles') ? undefined : json['roles'],
        'userEmail': !exists(json, 'userEmail') ? undefined : json['userEmail'],
        'activeDirectoryGroups': !exists(json, 'activeDirectoryGroups') ? undefined : ((json['activeDirectoryGroups'] as Array<any>).map(GetProjectAssignmentsResponseItemADGroupFromJSON)),
        'activeDirectoryGroupStatus': !exists(json, 'activeDirectoryGroupStatus') ? undefined : json['activeDirectoryGroupStatus'],
    };
}

export function GetProjectAssignmentsResponseItemToJSON(value?: GetProjectAssignmentsResponseItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'userId': value.userId,
        'roles': value.roles,
        'userEmail': value.userEmail,
        'activeDirectoryGroups': value.activeDirectoryGroups === undefined ? undefined : ((value.activeDirectoryGroups as Array<any>).map(GetProjectAssignmentsResponseItemADGroupToJSON)),
        'activeDirectoryGroupStatus': value.activeDirectoryGroupStatus,
    };
}